requirements.txt
distro==1.4.0
numpy==1.17.0
pandas==0.25.0
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
tabula-py==1.4.1